requirements.txt
# Deep Learning - PyTorch (for inference with existing .pth model)
torch>=2.0.0
torchvision>=0.15.0

# Optional: TensorFlow (only if you want to train new models)
# tensorflow>=2.13.0
# keras>=2.13.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization (for training notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
scikit-learn>=1.3.0